added cache checking
